+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - new helper routeBack to get url of parent route
 - plugin admin close panel
 - increased coverage for jsdoc comments in the code
+- plugin admin new option K.settings.admin.emailOnStartup
 
 ## v1.6.0
 - rename base package from ui to core-ui